requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-socketio==5.11.4

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9

# Config
pydantic-settings==2.6.1
python-dotenv==1.0.1

# SSH
asyncssh==2.18.0

# AI
anthropic>=0.43.0
openai>=1.40.0

# STIX
stix2==3.0.2

# HTTP client (MCP)
httpx==0.28.1

# Vault
hvac==2.3.0

# Utils
python-multipart==0.0.20
aiofiles==24.1.0
pyyaml==6.0.2
